pyproject.toml
[project]
name = "bb-troubleshooter-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2024.11.0",
    "ipykernel>=6.30.1",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.8",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.2",
    "streamlit>=1.50.0",
    "structlog>=25.4.0",
    "tabulate>=0.9.0",
]
